Revert platform specific sizing
